move common functions to base class
